Removed unecessary code
