+++ backend/authentication/decorators.py
@@ -76,7 +76,6 @@ def roles_required(*required_rolenames):
 def user_session_exists(f):
     @wraps(f)
     def wrap(*args, **kwargs):
-        print(session)
         if "profile" in session:
             return f(*args, **kwargs)
         else:
